requirements.txt
python-dateutil>=2.9.0.post0
requests>=2.32.3
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.1
openai>=1.52.1
tiktoken>=0.8.0